debug for unstable pressure readings

+++ esphome/components/ams5915/ams5915.h
@@ -52,7 +52,7 @@ class Ams5915 : public PollingComponent, public sensor::Sensor, public i2c::I2CD
     Transducer _type;
   protected:
 
-  std::string intToBinary(int n);
+  std::string intToBinary(uint16_t n);
   // struct to hold sensor data
     struct Data {
       float Pressure_Pa;
